code refactoring in Matches.jsx
